Finished incident synchronisation
