+++ requirements.txt
@@ -1,7 +1,7 @@
 PyYAML>=5.4
-pyzabbix==1.2.1
+pyzabbix==1.3.1
 requests>=2.21.0
 pytz==2024.1
 # Dev only
-pytest==8.1.1
+pytest==8.2.0
 pytest-env==1.1.3
